# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sabah Park Api Documentation",
        "contact": {
            "email": "prastiyo.beka12@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/dashboard/summary": {
            "get": {
                "tags": [
                    "DASHBOARD"
                ],
                "summary": "Get Data Member",
                "operationId": "App\\Http\\Controllers\\Admin\\DashboardController::summary",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/admin/destination/type": {
            "get": {
                "tags": [
                    "Setup Destination (BackOffice)"
                ],
                "summary": "Get Type Destination",
                "operationId": "App\\Http\\Controllers\\DestinationController::get_type",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/admin/destination/get": {
            "get": {
                "tags": [
                    "Setup Destination (BackOffice)"
                ],
                "summary": "Get Data Destination",
                "operationId": "App\\Http\\Controllers\\DestinationController::get",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/admin/destination/manage": {
            "post": {
                "tags": [
                    "Setup Destination (BackOffice)"
                ],
                "summary": "Manage Destination",
                "operationId": "App\\Http\\Controllers\\DestinationController::manage",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "destid": {
                                        "type": "string"
                                    },
                                    "destname": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "desttype": {
                                        "type": "string"
                                    },
                                    "loc_price_18above": {
                                        "type": "string"
                                    },
                                    "loc_price_18below": {
                                        "type": "string"
                                    },
                                    "int_price_18above": {
                                        "type": "string"
                                    },
                                    "int_price_18below": {
                                        "type": "string"
                                    },
                                    "loc_quota": {
                                        "type": "string"
                                    },
                                    "int_quota": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/admin/destination/delete": {
            "post": {
                "tags": [
                    "Setup Destination (BackOffice)"
                ],
                "summary": "Delete Destination",
                "operationId": "App\\Http\\Controllers\\DestinationController::delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "destid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/public/order/destination/id": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Get Destination",
                "operationId": "App\\Http\\Controllers\\DestinationTypeController::get",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian Judul",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id_type",
                        "in": "query",
                        "description": "Filter Berdasarkan Type Destination",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "index start data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": " data yang ingin di tampilkan",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/member/order/destination/id": {
            "get": {
                "tags": [
                    "ORDER"
                ],
                "summary": "Get Destination",
                "operationId": "App\\Http\\Controllers\\DestinationTypeController::by_id",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "destid",
                        "in": "query",
                        "description": "ID DESTINATION",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian Judul",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id_type",
                        "in": "query",
                        "description": "Filter Berdasarkan Type Destination",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "index start data",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": " data yang ingin di tampilkan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/member/order/destination": {
            "get": {
                "tags": [
                    "ORDER"
                ],
                "summary": "Get Destination",
                "operationId": "App\\Http\\Controllers\\OrderController::get",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian Judul",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id_type",
                        "in": "query",
                        "description": "Filter Berdasarkan Type Destination",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "index start data",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": " data yang ingin di tampilkan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}